Refactor performance code for better organization
