+++ packages/devtools_app/lib/devtools_app.dart
@@ -39,10 +39,10 @@ export 'src/screens/network/network_controller.dart';
 export 'src/screens/network/network_model.dart';
 export 'src/screens/network/network_screen.dart';
 export 'src/screens/performance/panes/controls/enhance_tracing/enhance_tracing_controller.dart';
+export 'src/screens/performance/panes/raster_metrics/raster_metrics_controller.dart';
 export 'src/screens/performance/performance_controller.dart';
 export 'src/screens/performance/performance_model.dart';
 export 'src/screens/performance/performance_screen.dart';
-export 'src/screens/performance/raster_metrics_controller.dart';
 export 'src/screens/performance/timeline_streams.dart';
 export 'src/screens/profiler/cpu_profile_model.dart';
 export 'src/screens/profiler/profile_granularity.dart';
