+++ packages/devtools_app/lib/src/screens/performance/panes/timeline_events/timeline_flame_chart.dart
@@ -8,21 +8,21 @@ import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-import '../../charts/flame_chart.dart';
-import '../../primitives/auto_dispose_mixin.dart';
-import '../../primitives/flutter_widgets/linked_scroll_controller.dart';
-import '../../primitives/geometry.dart';
-import '../../primitives/trace_event.dart';
-import '../../primitives/utils.dart';
-import '../../shared/common_widgets.dart';
-import '../../shared/notifications.dart';
-import '../../shared/theme.dart';
-import '../../shared/utils.dart';
-import '../../ui/colors.dart';
-import '../../ui/utils.dart';
-import 'performance_controller.dart';
-import 'performance_model.dart';
-import 'performance_utils.dart';
+import '../../../../charts/flame_chart.dart';
+import '../../../../primitives/auto_dispose_mixin.dart';
+import '../../../../primitives/flutter_widgets/linked_scroll_controller.dart';
+import '../../../../primitives/geometry.dart';
+import '../../../../primitives/trace_event.dart';
+import '../../../../primitives/utils.dart';
+import '../../../../shared/common_widgets.dart';
+import '../../../../shared/notifications.dart';
+import '../../../../shared/theme.dart';
+import '../../../../shared/utils.dart';
+import '../../../../ui/colors.dart';
+import '../../../../ui/utils.dart';
+import '../../performance_controller.dart';
+import '../../performance_model.dart';
+import '../../performance_utils.dart';
 
 class TimelineEventsView extends StatelessWidget {
   const TimelineEventsView({
