+++ packages/devtools_app/lib/src/screens/performance/performance_controller.dart
@@ -28,10 +28,10 @@ import '../profiler/cpu_profile_service.dart';
 import '../profiler/cpu_profile_transformer.dart';
 import '../profiler/profile_granularity.dart';
 import 'panes/controls/enhance_tracing/enhance_tracing_controller.dart';
+import 'panes/raster_metrics/raster_metrics_controller.dart';
 import 'performance_model.dart';
 import 'performance_screen.dart';
 import 'performance_utils.dart';
-import 'raster_metrics_controller.dart';
 import 'rebuild_counts.dart';
 import 'simple_trace_example.dart';
 import 'timeline_event_processor.dart';
