+++ packages/devtools_app/lib/src/screens/performance/tabbed_performance_view.dart
@@ -13,12 +13,12 @@ import '../../shared/globals.dart';
 import '../../shared/theme.dart';
 import '../../ui/search.dart';
 import '../../ui/tab.dart';
-import 'frame_analysis.dart';
+import 'panes/frame_analysis/frame_analysis.dart';
+import 'panes/raster_metrics/raster_metrics.dart';
+import 'panes/timeline_events/timeline_flame_chart.dart';
 import 'performance_controller.dart';
 import 'performance_model.dart';
 import 'performance_screen.dart';
-import 'raster_metrics.dart';
-import 'timeline_flame_chart.dart';
 
 final timelineSearchFieldKey = GlobalKey(debugLabel: 'TimelineSearchFieldKey');
 
@@ -181,6 +181,27 @@ class _TabbedPerformanceViewState extends State<TabbedPerformanceView>
       tabName: tabName,
       gaPrefix: _gaPrefix,
       trailing: trailing,
+    );
+  }
+}
+
+@visibleForTesting
+class RefreshTimelineEventsButton extends StatelessWidget {
+  const RefreshTimelineEventsButton({
+    Key? key,
+    required this.controller,
+  }) : super(key: key);
+
+  final PerformanceController controller;
+
+  @override
+  Widget build(BuildContext context) {
+    return DevToolsIconButton(
+      iconData: Icons.refresh,
+      onPressed: controller.processAvailableEvents,
+      tooltip: 'Refresh timeline events',
+      gaScreen: analytics_constants.performance,
+      gaSelection: analytics_constants.refreshTimelineEvents,
     );
   }
 }
