+++ packages/devtools_app/test/performance/tabbed_performance_view_test.dart
@@ -6,12 +6,12 @@ import 'package:devtools_app/src/charts/flame_chart.dart';
 import 'package:devtools_app/src/config_specific/ide_theme/ide_theme.dart';
 import 'package:devtools_app/src/config_specific/import_export/import_export.dart';
 import 'package:devtools_app/src/primitives/listenable.dart';
-import 'package:devtools_app/src/screens/performance/frame_analysis.dart';
+import 'package:devtools_app/src/screens/performance/panes/frame_analysis/frame_analysis.dart';
+import 'package:devtools_app/src/screens/performance/panes/raster_metrics/raster_metrics.dart';
+import 'package:devtools_app/src/screens/performance/panes/timeline_events/timeline_flame_chart.dart';
 import 'package:devtools_app/src/screens/performance/performance_controller.dart';
 import 'package:devtools_app/src/screens/performance/performance_model.dart';
-import 'package:devtools_app/src/screens/performance/raster_metrics.dart';
 import 'package:devtools_app/src/screens/performance/tabbed_performance_view.dart';
-import 'package:devtools_app/src/screens/performance/timeline_flame_chart.dart';
 import 'package:devtools_app/src/service/service_manager.dart';
 import 'package:devtools_app/src/shared/common_widgets.dart';
 import 'package:devtools_app/src/shared/globals.dart';
